requirements.txt
joblib==1.2.0
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
scikit_learn==1.2.2
seaborn==0.12.2
streamlit==1.19.0
sklearn2==0.0.13